% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limonade.R
\name{limonade}
\alias{limonade}
\title{limonade}
\usage{
limonade(filename, format)
}
\arguments{
\item{filename}{filename}

\item{format}{format}
}
\description{
limonade
}
\examples{
\dontrun{
limonade()
}
}
\keyword{limonade}

